fix: resolve code review findings for GitLab MR !2
Reviewed with parallel toolkit agents + codex/gemini plus manual review of
the crypto core, WASM boundary, and account state machine.

Correctness / security (fixed):
- shrincs_verifier_types: remove the unconditional top-level parameter
  constants that silently shadowed `pub use profile::*`. An explicit item
  takes precedence over a glob re-export, so profile-128s-q18/q20 builds
  were using 256s hypertree/FORS counts and the wrong stateless budget.
  The default 256s build is byte-identical (values coincided). [codex P1]
- shrincs_verifier_fors_c: fail closed on the hypertree-coordinate bit
  reads (unwrap_or(0) -> ?), matching the signer and the rest of the
  verifier. Not exploitable under current constants; removes a future
  footgun. [silent-failure-hunter]
- shrincs_verifier_hypertree: guard subtree_height against u32::BITS (not
  u64::BITS) before the u32 shift; add the matching guard to the signer.
- account: use saturating_add for nextStatefulLeafIndex (no release
  overflow-checks; verifier already caps usable leaves).

Performance (fixed):
- shrincs_signer_stateful: drop the redundant stateful_auth_path rebuild;
  sign_stateful_raw_at_leaf already produced the identical path, so normal
  signing paid the dominant cost twice. [codex P2]

Config / CI (fixed):
- Cargo.toml: pin wasm-bindgen to =0.2.100 (must equal the CLI pin).
- .gitlab-ci.yml: add --locked to cargo build/test, gate publish-npm to
  semver tags, pin rust:1.89.0 image (matches rust-toolchain.toml).

TypeScript (fixed):
- loader.browser.ts: cache the in-flight init promise so concurrent
  callers share one WebAssembly.instantiate; clear on reject to allow retry.

Cleanup / tests (fixed):
- Zero clippy warnings: div_ceil, ptr_arg, needless_borrow, unnecessary
  cast, Default for ShrincsVerifier, unsafe-in-name alias, dead parse_hex_bytes
  gated to cfg(test); too_many_arguments allowed with justification on the
  ADRS-positional signer helpers.
- wotsplus: test_invalid_signature_length now exercises the real length
  guard (was building a correct-length signature).
- account: add increment_u256_be byte-carry test.
- signer: use the named ADDRESS_TYPE_WOTS_HASH constant.

Dismissed: identical wots-c-chain domain tag stateless/stateful (Solidity
parity; separation via pk_seed), signer Option->thiserror (invasive), super::
imports (idiomatic), ignored solidity_account_vectors (vectors intentionally
gitignored, copied manually). Noted as follow-ups: consume the committed 256s
KAT in a Rust test, run the WASM account bindings via Node conformance.

Verified: cargo clippy (0 warnings), cargo test --features wasm-bindings
(65 lib + integration KAT pass), wasm32 release build (--locked), default
and profile-128s-q18 builds.

+++ src/shrincs/shrincs_verifier_hypertree.rs
@@ -45,7 +45,10 @@ pub(crate) fn verify_hypertree(
     if subtree_height == 0 {
         return false;
     }
-    if subtree_height >= u64::BITS {
+    // `leaf_count` below shifts a u32, so the guard must be u32::BITS (not
+    // u64::BITS): a profile with subtree_height in 32..64 would otherwise pass
+    // this check and panic on the `1u32 << subtree_height` shift.
+    if subtree_height >= u32::BITS {
         return false;
     }
     let leaf_count = 1u32 << subtree_height;
